Refactor dotfiles layout and add validation CI
