+++ git/gitconfig
@@ -1,0 +1,24 @@
+[user]
+    email = eseliavka@example.com
+    name = Evgenii Seliavka
+[alias]
+    st = status
+    co = checkout
+    br = branch
+    up = rebase
+    ci = commit
+    lg = log --color --graph --pretty=format:'%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit
+    spush = push --recurse-submodules=on-demand
+    pl = pull --quiet
+[core]
+    editor = vim
+    autocrlf = input
+[status]
+    submoduleSummary = true
+[diff]
+    submodule = log
+    renames = true
+[color]
+    ui = true
+[include]
+    path = ~/.gitconfig.local
