+++ git/gitconfig.local.example
@@ -1,0 +1,5 @@
+[user]
+    email = you@example.com
+
+[core]
+    sshCommand = ssh -i ~/.ssh/id_ed25519
