+++ psql/queries/storage.sql
@@ -1,0 +1,13 @@
+\set rsize 'SELECT nspname || \'.\' || relname AS "relation" ,pg_size_pretty(pg_total_relation_size(C.oid)) AS "total_size" FROM pg_class C LEFT JOIN pg_namespace N ON (N.oid = C.relnamespace) WHERE nspname NOT IN ( \'pg_catalog\' ,\'information_schema\') AND C.relkind <> \'i\' AND nspname !~ \'^pg_toast\' ORDER BY pg_total_relation_size(C.oid) DESC LIMIT 20;'
+
+\set tblbloat 'SELECT current_database()  ,schemaname  ,tblname  ,bs * tblpages AS real_size  ,(tblpages - est_num_pages) * bs AS bloat_size  ,tblpages  ,is_na  ,CASE   WHEN tblpages - est_num_pages > 0    THEN 100 * (tblpages - est_num_pages) / tblpages::FLOAT   ELSE 0   END AS bloat_ratio FROM (  SELECT ceil(reltuples / ((bs - page_hdr) / tpl_size)) + ceil(toasttuples / 4) AS est_num_pages   ,tblpages   ,bs   ,tblid   ,schemaname   ,tblname   ,heappages   ,toastpages   ,is_na  FROM (   SELECT (     4 + tpl_hdr_size + tpl_data_size + (2 * ma) - CASE      WHEN tpl_hdr_size % ma = 0       THEN ma      ELSE tpl_hdr_size % ma      END - CASE      WHEN ceil(tpl_data_size)::INT % ma = 0       THEN ma      ELSE ceil(tpl_data_size)::INT % ma      END     ) AS tpl_size    ,bs - page_hdr AS size_per_block    ,(heappages + toastpages) AS tblpages    ,heappages    ,toastpages    ,reltuples    ,toasttuples    ,bs    ,page_hdr    ,tblid    ,schemaname    ,tblname    ,is_na   FROM (    SELECT tbl.oid AS tblid     ,ns.nspname AS schemaname     ,tbl.relname AS tblname     ,tbl.reltuples     ,tbl.relpages AS heappages     ,coalesce(toast.relpages, 0) AS toastpages     ,coalesce(toast.reltuples, 0) AS toasttuples     ,current_setting(\'block_size\')::NUMERIC AS bs     ,CASE      WHEN version() ~ \'mingw32\'       OR version() ~ \'64-bit|x86_64|ppc64|ia64|amd64\'       THEN 8      ELSE 4      END AS ma     ,24 AS page_hdr     ,23 + CASE      WHEN MAX(coalesce(null_frac, 0)) > 0       THEN (7 + count(*)) / 8      ELSE 0::INT      END + CASE      WHEN tbl.relhasoids       THEN 4      ELSE 0      END AS tpl_hdr_size     ,sum((1 - coalesce(s.null_frac, 0)) * coalesce(CASE        WHEN t.typlen = - 1         THEN CASE           WHEN s.avg_width < 127            THEN s.avg_width + 1           ELSE s.avg_width + 4           END        WHEN t.typlen = - 2         THEN s.avg_width + 1        ELSE t.typlen        END, 1024)) AS tpl_data_size     ,bool_or(att.atttypid = \'pg_catalog.name\'::regtype) AS is_na    FROM pg_attribute AS att    JOIN pg_type AS t ON att.atttypid = t.oid    JOIN pg_class AS tbl ON att.attrelid = tbl.oid    JOIN pg_namespace AS ns ON ns.oid = tbl.relnamespace    JOIN pg_stats AS s ON s.schemaname = ns.nspname     AND s.tablename = tbl.relname     AND s.inherited = false     AND s.attname = att.attname    LEFT JOIN pg_class AS toast ON tbl.reltoastrelid = toast.oid    WHERE att.attnum > 0     AND NOT att.attisdropped     AND tbl.relkind = \'r\'     AND ns.nspname NOT IN (      \'information_schema\'      ,\'pg_catalog\'      )    GROUP BY 1     ,2     ,3     ,4     ,5     ,6     ,7     ,8     ,9     ,10     ,tbl.relhasoids    ORDER BY 2     ,3    ) AS s   ) AS s2  ) AS s3;'
+
+\set tblstorageprop 'SELECT n.nspname  ,c.relname  ,pg_catalog.array_to_string(c.reloptions || array(SELECT \'toast.\' || x FROM pg_catalog.unnest(tc.reloptions) x), \', \') AS relopts FROM pg_catalog.pg_class c LEFT JOIN pg_catalog.pg_class tc ON (c.reltoastrelid = tc.oid) JOIN pg_namespace n ON c.relnamespace = n.oid  AND n.nspname NOT IN (   \'information_schema\'   ,\'pg_catalog\'   ) WHERE c.relkind = \'r\';'
+
+\set tblage 'SELECT c.oid::regclass AS table_name  ,greatest(age(c.relfrozenxid), age(t.relfrozenxid)) AS age FROM pg_class c JOIN pg_namespace n ON c.relnamespace = n.oid  AND n.nspname NOT IN (   \'information_schema\'   ,\'pg_catalog\'   ) LEFT JOIN pg_class t ON c.reltoastrelid = t.oid WHERE c.relkind = \'r\';'
+
+\set vacuuminfo 'SELECT datname, age(datfrozenxid) FROM pg_database;'
+
+\set uselesscol 'SELECT nspname  ,relname  ,attname  ,typname  ,(stanullfrac * 100)::INT AS null_percent  ,CASE   WHEN stadistinct >= 0    THEN stadistinct   ELSE abs(stadistinct) * reltuples   END AS "distinct"  ,CASE 1   WHEN stakind1    THEN stavalues1   WHEN stakind2    THEN stavalues2   END AS "values" FROM pg_class c JOIN pg_namespace ns ON (ns.oid = relnamespace) JOIN pg_attribute ON (c.oid = attrelid) JOIN pg_type t ON (t.oid = atttypid) JOIN pg_statistic ON (   c.oid = starelid   AND staattnum = attnum   ) WHERE nspname NOT IN ('information_schema', 'pg_catalog')  AND relkind = 'r'  AND NOT attisdropped  AND attstattarget != 0  AND reltuples >= 100  AND stadistinct BETWEEN 0   AND 1 ORDER BY nspname  ,relname  ,attname;'
+
+\set avneeded 'SELECT *  ,n_dead_tup > av_threshold AS "av_needed"  ,CASE   WHEN reltuples > 0    THEN round(100.0 * n_dead_tup / (reltuples))   ELSE 0   END AS pct_dead FROM (  SELECT N.nspname   ,C.relname   ,pg_stat_get_tuples_inserted(C.oid) AS n_tup_ins   ,pg_stat_get_tuples_updated(C.oid) AS n_tup_upd   ,pg_stat_get_tuples_deleted(C.oid) AS n_tup_del   ,pg_stat_get_live_tuples(C.oid) AS n_live_tup   ,pg_stat_get_dead_tuples(C.oid) AS n_dead_tup   ,C.reltuples AS reltuples   ,round(current_setting(\'autovacuum_vacuum_threshold\')::INTEGER + current_setting(\'autovacuum_vacuum_scale_factor\')::NUMERIC * C.reltuples) AS av_threshold   ,date_trunc(\'minute\', greatest(pg_stat_get_last_vacuum_time(C.oid), pg_stat_get_last_autovacuum_time(C.oid))) AS last_vacuum   ,date_trunc(\'minute\', greatest(pg_stat_get_last_analyze_time(C.oid), pg_stat_get_last_analyze_time(C.oid))) AS last_analyze  FROM pg_class C  LEFT JOIN pg_namespace N ON (N.oid = C.relnamespace)  WHERE C.relkind IN (    \'r\'    ,\'t\'    )   AND N.nspname NOT IN (    \'pg_catalog\'    ,\'information_schema\'    )   AND N.nspname !~ \'^pg_toast\'  ) AS av ORDER BY av_threshold DESC, av_needed DESC  ,n_dead_tup DESC;'
