+++ shell/env.sh
@@ -1,0 +1,3 @@
+# shellcheck shell=bash
+export PAGER=less
+export LESS="-IMSx4 -FJXR"
