+++ shell/prompt.sh
@@ -1,0 +1,3 @@
+# shellcheck shell=bash
+PS1='[\t][\j][\u@\h:\W]\$ '
+export PS1
